adds complete MyTinyToDo integration test
