+++ .github/workflows/maven.yml
@@ -19,6 +19,19 @@ jobs:
         image: opentosca/ode:latest
         ports:
           - 9763:9763
+      engine-ia:
+        image: opentosca/engine-ia:latest
+        ports:
+          - '8090:8080'
+      dind:
+        image: jpetazzo/dind:latest
+        privileged: true
+        environment:
+          - PORT=2375
+          - DOCKER_DAEMON_ARGS=--dns 8.8.8.8 --dns 8.8.4.4 -D
+        ports:
+          - '2375:2375'
+          - '9990-9999:9990-9999'
 
     steps:
     - uses: actions/checkout@v2
