+++ org.opentosca.bus/org.opentosca.bus.application.service/src/main/java/org/opentosca/bus/application/service/impl/servicehandler/ApplicationBusPluginRegistry.java
@@ -1,6 +1,7 @@
 package org.opentosca.bus.application.service.impl.servicehandler;
 
 import java.util.Collection;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
 import javax.inject.Inject;
@@ -29,9 +30,9 @@ public class ApplicationBusPluginRegistry {
     final private static Logger LOG = LoggerFactory.getLogger(ApplicationBusPluginRegistry.class);
 
     @Inject
-    public ApplicationBusPluginRegistry(@Autowired(required = false) Collection<IApplicationBusPluginService> plugins) {
-        if (plugins != null) {
-            for (IApplicationBusPluginService plugin : plugins) {
+    public ApplicationBusPluginRegistry(@Autowired(required = false) Optional<Collection<IApplicationBusPluginService>> plugins) {
+        if (plugins != null && plugins.isPresent()) {
+            for (IApplicationBusPluginService plugin : plugins.get()) {
                 final String routingEndpoint = plugin.getRoutingEndpoint();
                 plugin.getSupportedInvocationTypes().forEach(invocationType -> pluginServices.put(invocationType, routingEndpoint));
             }
