+++ pom.xml
@@ -16,11 +16,13 @@
 		<camel.version>2.18.5</camel.version>
 		<commons-io.version>2.6</commons-io.version>
 		<jaxb.version>2.2.11</jaxb.version>
+        <junit.version>4.13.2</junit.version>
+        <logback.version>1.1.11</logback.version>
 		<!-- needs to match the jersey version used in winery to avoid conflicts -->
 		<jersey.version>2.30.1</jersey.version>
-		<spring.version>4.3.18.RELEASE</spring.version>
+		<spring.version>4.3.25.RELEASE</spring.version>
 		<swagger.version>1.5.24</swagger.version>
-		<resteasy.version>3.1.4.Final</resteasy.version>
+        <tomcat.version>9.0.43</tomcat.version>
 	</properties>
 
 	<repositories>
@@ -115,13 +117,13 @@
 		<dependency>
 			<groupId>ch.qos.logback</groupId>
 			<artifactId>logback-classic</artifactId>
-			<version>1.1.2</version>
+			<version>${logback.version}</version>
 		</dependency>
 
 		<dependency>
 			<groupId>ch.qos.logback</groupId>
 			<artifactId>logback-core</artifactId>
-			<version>1.1.2</version>
+			<version>${logback.version}</version>
 		</dependency>
 
 		<dependency>
@@ -130,10 +132,11 @@
 			<version>2.1.0</version>
 			<scope>compile</scope>
 		</dependency>
+
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
+            <version>${junit.version}</version>
             <scope>test</scope>
         </dependency>
